akicti/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-redis>=5.4.0
django-ratelimit>=4.1.0
django-filter>=23.0
python-decouple>=3.8
psycopg2-binary>=2.9.9
djoser>=2.3.0
gunicorn>=21.2.0
whitenoise>=6.6.0
